Store position attribute as Array type
The list of position attributes has now been replaced with a single
attribute, which can be an Array. Postgre will store the array directly,
SQLite will create one column for each element.

This commit also tightens the requirements for optional arguments
passed to the graph databases and simplifies the meta-data checks.

There is now also documentation on the allowed read/write modes and
checks for the passed arguments.