fix setter failing in github actions
dask arrays are not supposed to cache results from reads, so if you change the data that the dask array wraps you should get the updated data on the next read.
This does not seem to be happening for dask arrays wrapping in-memory numpy arrays on some hardware. It is consistent behavior in the github actions but I cannot recreate it locally.